Adds new request.$wheelsDebugSQL flag
This allows us to get the generated SQL from a DB migration without actually executing it, or incremented/deincrementing the migration schema version.

+++ wheels/migrator/functions.cfm
@@ -226,10 +226,11 @@ public string function redoMigration(string version="") {
  */
 private void function $setVersionAsMigrated(required string version) {
 	local.appKey = $appKey();
-	$query(
-		datasource=application[local.appKey].dataSourceName,
-		sql="INSERT INTO #application[local.appKey].migratorTableName# (version) VALUES ('#$sanitiseVersion(arguments.version)#')"
-	);
+	if(!structKeyExists(request, "$wheelsDebugSQL"))
+		$query(
+			datasource=application[local.appKey].dataSourceName,
+			sql="INSERT INTO #application[local.appKey].migratorTableName# (version) VALUES ('#$sanitiseVersion(arguments.version)#')"
+		);
 }
 
 /**
@@ -237,10 +238,11 @@ private void function $setVersionAsMigrated(required string version) {
  */
 private void function $removeVersionAsMigrated(required string version) {
 	local.appKey = $appKey();
-	$query(
-		datasource=application[local.appKey].dataSourceName,
-		sql="DELETE FROM #application[local.appKey].migratorTableName# WHERE version = '#$sanitiseVersion(arguments.version)#'"
-	);
+	if(!structKeyExists(request, "$wheelsDebugSQL"))
+		$query(
+			datasource=application[local.appKey].dataSourceName,
+			sql="DELETE FROM #application[local.appKey].migratorTableName# WHERE version = '#$sanitiseVersion(arguments.version)#'"
+		);
 }
 
 /**
